refactor(cxo): extract cxosub; wire standalone reward server to CXO feeds
Extract the visor's intermittent CXO subscription manager
(pkg/visor/cxo_subscription_manager.go, ~800 lines) into a new lightweight,
reusable package pkg/cxo/cxosub, decoupled from *Visor via an injected
Deps{Dmsg, FeedSpec, Log}. pkg/visor keeps working unchanged via type
aliases + a thin cxoFeedSpec shim; its netviz behavior and tests are
preserved.

Wire the STANDALONE reward server (tp-viz serving theskywirenetwork.net
over dmsg) to the same intermittent subscriber instead of HTTP-over-dmsg
polling: tpviz.SetCXOSubMgrFromDmsg builds a cxosub.Manager over the
server's dmsg client, resolving TPD/SD/DMSG-D feed publishers from the
config's dmsg:// service URLs (cxosub.FeedRoute centralizes the feed →
port/prefix mapping). handleUptimes + handleTransports try CXO first
(tryCXOUptimes/tryCXOTransports) and fall back to the HTTP-over-dmsg path
on a cache miss; services + dmsg-clients already had CXO walks that now
activate.

Also make the first-Root wait per-feed (cxosub.FeedFirstSyncTimeout): the
all-transports snapshot (~10MB, the most important feed for the graph)
intermittently exceeded the flat 10s bound and failed to sync over CXO;
it now gets 45s, other feeds keep 10s. Threaded through both the cycle and
the RPC fetch/refresh default.

Fixes the transport-graph showing ~9 of ~1011 visors (empty /api/uptimes
-> all 'unknown' -> hidden) via the correct mechanism: the TPD-integrated
feeds over CXO, intermittent subscribe->snapshot->unsubscribe (self-heals
across TPD restarts, no persistent subscriber state). Verified live: TPD
uptime feed publishes 1084 visors / 857 online. Supersedes #3388's
HTTP-over-dmsg patch (kept as fallback). Design: docs/tpviz-cxo-subscriber.md.
